+++ whatever/src/app/page.tsx
@@ -1,11 +1,10 @@
 export default function Home() {
   return (
-    // <section className='min-h-screen mt-4/5'>
-    //   <div className='flex bg-custom-image min-h-full bg-cover bg-center'>
-    //     <div className='absolute w-full h-full inset-0 bg-black bg-opacity-50' />
-    //     hey
-    //   </div>
-    // </section>
-    <div />
+    <section className='min-h-screen mt-4/5'>
+      {/* <div className='flex bg-custom-image min-h-full bg-cover bg-center'>
+        <div className='absolute w-full h-full inset-0 bg-black bg-opacity-50' />
+        hey
+      </div> */}
+    </section>
   );
 }
